refactor: split puzzle helpers
